smart_driver/requirements.txt
dj-database-url==0.4.1
Django==1.9.6
djangorestframework==3.3.3
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==3.2
